nlp/pyproject.toml
[project]
name = "nlp"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "ingredient-parser-nlp (>=1.3.2,<2.0.0)",
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "recipe-scrapers (>=15.4.0,<16.0.0)",
    "inflect (>=7.5.0,<8.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"